QuarkFspPkg: FSP 1.1 fix Visual Studio build errors
Fix build errors detected with Visual Studio 2013.

Contributed-under: TianoCore Contribution Agreement 1.0

+++ Library/Fsp1_1Lib/Fsp1_1Lib.c
@@ -142,7 +142,7 @@ VOID GetMemoryParameters (
   MrcData->channel_width       = MemoryInitUpd->ChanWidth;
   MrcData->address_mode        = MemoryInitUpd->AddrMode;
   // Enable scrambling if requested.
-  MrcData->scrambling_enables  = MemoryInitUpd->Flags;
+  MrcData->scrambling_enables  = (UINT8)MemoryInitUpd->Flags;
   MrcData->ddr_type            = MemoryInitUpd->DramType;
   MrcData->dram_width          = MemoryInitUpd->DramWidth;
   MrcData->ddr_speed           = MemoryInitUpd->DramSpeed;
